+++ Front/covoit-viveris/src/components/booking/BookCarshareView.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import * as levels from "../../functions/levels";
-import { BookCarShareProfile } from "./BookCarShareProfile";
+import { BookCarshareProfile } from "./BookCarshareProfile.jsx";
 
-export function BookCarShareView(userDriver){
+export function BookCarshareView(userDriver){
 
     const [ownedBadges, setOwnedBadges] = useState([]);
 
@@ -12,7 +12,6 @@ export function BookCarShareView(userDriver){
             return res.json();
         })
         .then((data) => {
-            console.log(data)
             setOwnedBadges(data);
         });
     }, []);
@@ -27,7 +26,7 @@ export function BookCarShareView(userDriver){
                 </div>
                 <div className="row">
                     <div className="col">
-                        <BookCarShareProfile userid={userDriver.carDriver.uid}/>
+                        <BookCarshareProfile userid={userDriver.carDriver.uid}/>
                     </div>
                     <div className="col" style={{ maxWidth:"10px" }}>
                         <div className="line-company-color"></div>
