+++ Front/covoit-viveris/src/components/research/CarshareResearchProfile.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 
-export function BookCarShareProfile(userid){
+export function CarshareResearchProfile(userid){
 
     const [userDriver, setUserDriver] = useState();
 
@@ -10,7 +10,6 @@ export function BookCarShareProfile(userid){
             return res.json();
           })
           .then((data) => {
-            console.log(data)
             setUserDriver(data);
           });
     }, []);
@@ -19,9 +18,12 @@ export function BookCarShareProfile(userid){
         <React.Fragment>
             {userDriver &&
                 <div>
-                    <div className="row" style={{ marginTop: "-145px"}}>
-                        <img className="center-picture" src="../../src/images/profil_picture.png" alt="Photo profil" style={{ width:"200px" }}/>
-                        <div style={{ marginLeft:"45px", marginTop:"-30px"}}>
+                    <div className="row" style={{ marginTop: "25px" }}>
+                        <div className="col center-picture" style={{ maxWidth:"160px" }}>
+                            <img src="./src/images/profil_picture.png" alt="Photo profil" width="150px"/>
+                            {/* <p className={`color-company ${marginClass}`} style={{ marginTop: "-58px" }}><strong>{u.user.experience}</strong></p> */}
+                        </div>
+                        <div className="col">
                             <p style={{ marginTop: "15px", fontSize: "20px" }}><strong>{userDriver.pseudo}</strong></p>
                             <p style={{ fontSize:  "14px" }}>{userDriver.job}, {userDriver.city}</p>
                             <p><em style={{ fontSize:  "14px" }}>{userDriver.nb_carshares} covoiturages, {userDriver.kilometers} km parcourus</em></p>
@@ -29,6 +31,8 @@ export function BookCarShareProfile(userid){
                     </div>
                 </div>
             }
+            {/* <img className="center-picture" src="./src/images/profil_picture.png" alt="Photo profil" style={{ width: "250px", marginTop: "-150px" }}/> */}
+
         </React.Fragment>
     );
 }
