+++ Front/covoit-viveris/src/pages/Schedule.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Navbar } from "../components/header/Navbar";
-import CarshareCreation from "../components/carsharecreation/CarshareCreation.jsx";
+import CarshareCreation from "../components/creation/CarshareCreation.jsx";
 
 
 const Schedule = () => {
